Adjusted According to C++ Template

+++ Number_Theory/Prime_Factorization/__src/numOfDivisor.cpp
@@ -1,12 +1,12 @@
-long long numOfDivisor (long long n) {
+int64 numOfDivisor (int64 n) {
     if (n < 1LL) return 0LL;
-    long long cnt = 0LL;
-    long long ret = 1LL;
+    int64 cnt = 0LL;
+    int64 ret = 1LL;
 
     while (!(n & 1LL)) { n >>= 1LL; ++cnt; }
     ret *= (cnt + 1LL);
 
-    for (long long i = 3LL; i * i <= n; i += 2LL) {
+    for (int64 i = 3LL; i * i <= n; i += 2LL) {
         cnt = 0LL;
         while (!(n % i)) { n /= i; ++cnt; }
         ret *= (cnt + 1LL);
